requirements.txt
opencv-python
ultralytics
easyocr
transformers
requests
ccxt
python-telegram-bot
Pillow
telethon
python-dotenv
flask